+++ src/main/java/me/alzz/ext/Rx.kt
@@ -51,14 +51,14 @@ fun Disposable.disposeBy(d: DisposableContainer): Disposable {
 /**
  * 缓存结果
  */
-fun <T> Observable<T>.cache(ctx: Context, name: String, type: Class<T>, validDays: Int): Single<T> {
+fun <T> Observable<T>.cache(ctx: Context, name: String, type: Class<T>, validDays: Float): Single<T> {
     val cache = Cache.load(ctx, name, validDays, type).applySchedulers().ignoreError()
     val expensive = this
             .applySchedulers()
             .observeOn(Schedulers.io())
             .map { Cache.cache(ctx, name, it); it }
             .observeOn(AndroidSchedulers.mainThread())
-    val useCache = Cache.load(ctx, name, Int.MAX_VALUE, type).applySchedulers()
+    val useCache = Cache.load(ctx, name, Float.MAX_VALUE, type).applySchedulers()
 
     return Observable
             .concat(cache, expensive, useCache)
@@ -66,7 +66,7 @@ fun <T> Observable<T>.cache(ctx: Context, name: String, type: Class<T>, validDay
             .firstOrError()
 }
 
-fun <T> Observable<T>.cache(ctx: Context, name: String, validDays: Int, type: Type): Single<T> {
+fun <T> Observable<T>.cache(ctx: Context, name: String, validDays: Float, type: Type): Single<T> {
     val cache = Cache.load<T>(ctx, name, validDays, type).applySchedulers().ignoreError()
 
     val expensive = this
@@ -76,7 +76,7 @@ fun <T> Observable<T>.cache(ctx: Context, name: String, validDays: Int, type: Ty
         .observeOn(AndroidSchedulers.mainThread())
         .ignoreError()
 
-    val useCache = Cache.load<T>(ctx, name, Int.MAX_VALUE, type).applySchedulers()
+    val useCache = Cache.load<T>(ctx, name, Float.MAX_VALUE, type).applySchedulers()
 
     return Observable
         .concat(cache, expensive, useCache)
